refactor: consolidate movie-related utilities into moviesUtils
